Extracted all rules to common detectors
