+++ rules/ktlint/src/test/kotlin/com/twitter/rules/ktlint/compose/ComposeMutableParametersCheckTest.kt
@@ -4,6 +4,7 @@ package com.twitter.rules.ktlint.compose
 
 import com.pinterest.ktlint.test.KtLintAssertThat.Companion.assertThatRule
 import com.pinterest.ktlint.test.LintViolation
+import com.twitter.compose.rules.ComposeMutableParameters
 import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
 
@@ -29,22 +30,22 @@ class ComposeMutableParametersCheckTest {
             LintViolation(
                 line = 2,
                 col = 15,
-                detail = ComposeMutableParametersCheck.MutableParameterInCompose
+                detail = ComposeMutableParameters.MutableParameterInCompose
             ),
             LintViolation(
                 line = 4,
                 col = 15,
-                detail = ComposeMutableParametersCheck.MutableParameterInCompose
+                detail = ComposeMutableParameters.MutableParameterInCompose
             ),
             LintViolation(
                 line = 6,
                 col = 15,
-                detail = ComposeMutableParametersCheck.MutableParameterInCompose
+                detail = ComposeMutableParameters.MutableParameterInCompose
             ),
             LintViolation(
                 line = 8,
                 col = 15,
-                detail = ComposeMutableParametersCheck.MutableParameterInCompose
+                detail = ComposeMutableParameters.MutableParameterInCompose
             )
         )
     }
